added padding, sizeBoxes and AxisAlignment to fix UI design
